refactor(oracles/data_providers_sdk): Migrate exchange fetchers
