+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/binance_us.rs
@@ -1,40 +1,47 @@
 use anyhow::Result;
 use futures::{future::LocalBoxFuture, FutureExt};
 
+use itertools::Itertools;
 use serde::Deserialize;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
 
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
+
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-pub struct MEXCPriceData {
+#[serde(rename_all = "camelCase")]
+pub struct BinanceUsPriceData {
     pub symbol: String,
     #[serde(deserialize_with = "as_f64")]
-    #[serde(rename = "lastPrice")]
     pub last_price: f64,
     #[serde(deserialize_with = "as_f64")]
     pub volume: f64,
 }
 
-type MEXCPriceResponse = Vec<MEXCPriceData>;
+type BinanceUsPriceResponse = Vec<BinanceUsPriceData>;
 
-pub struct MEXCPriceFetcher;
+pub struct BinanceUsPriceFetcher<'a> {
+    pub symbols: &'a [String],
+}
 
-impl PricesFetcher<'_> for MEXCPriceFetcher {
-    const NAME: &'static str = "MEXC";
+impl<'a> PricesFetcher<'a> for BinanceUsPriceFetcher<'a> {
+    const NAME: &'static str = "Binance US";
 
-    fn new(_symbols: &[String], _api_key: Option<&str>) -> Self {
-        Self
+    fn new(symbols: &'a [String], _api_key: Option<&'a str>) -> Self {
+        Self { symbols }
     }
 
     fn fetch(&self) -> LocalBoxFuture<Result<PairPriceData>> {
         async {
-            let response = http_get_json::<MEXCPriceResponse>(
-                "https://api.mexc.com/api/v3/ticker/24hr",
-                None,
+            let req_symbols = format!(
+                "[{}]",
+                self.symbols.iter().map(|s| format!("\"{s}\"")).join(",")
+            );
+
+            let response = http_get_json::<BinanceUsPriceResponse>(
+                "https://api.binance.us/api/v3/ticker/24hr",
+                Some(&[("symbols", req_symbols.as_str())]),
                 None,
             )
             .await?;
