+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/bitget.rs
@@ -4,59 +4,52 @@ use futures::{future::LocalBoxFuture, FutureExt};
 use serde::Deserialize;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
 
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
+
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-pub struct CryptoComPriceData {
-    pub i: String,
+#[serde(rename_all = "camelCase")]
+pub struct BitgetPriceData {
+    pub symbol: String,
     #[serde(deserialize_with = "as_f64")]
-    pub a: f64,
+    pub close: f64,
     #[serde(deserialize_with = "as_f64")]
-    pub v: f64,
+    pub base_vol: f64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-pub struct CryptoComResult {
-    pub data: Vec<CryptoComPriceData>,
+#[serde(rename_all = "camelCase")]
+pub struct BitgetPriceResponse {
+    pub code: String,
+    pub data: Vec<BitgetPriceData>,
 }
+pub struct BitgetPriceFetcher;
 
-#[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-pub struct CryptoComPriceResponse {
-    pub code: i8,
-    pub result: CryptoComResult,
-}
-
-pub struct CryptoComPriceFetcher;
-
-impl PricesFetcher<'_> for CryptoComPriceFetcher {
-    const NAME: &'static str = "Crypto.com";
+impl PricesFetcher<'_> for BitgetPriceFetcher {
+    const NAME: &'static str = "Bitget";
 
     fn new(_symbols: &[String], _api_key: Option<&str>) -> Self {
         Self
     }
-
     fn fetch(&self) -> LocalBoxFuture<Result<PairPriceData>> {
         async {
-            let response = http_get_json::<CryptoComPriceResponse>(
-                "https://api.crypto.com/exchange/v1/public/get-tickers",
+            let response = http_get_json::<BitgetPriceResponse>(
+                "https://api.bitget.com/api/spot/v1/market/tickers",
                 None,
                 None,
             )
             .await?;
 
             Ok(response
-                .result
                 .data
                 .into_iter()
                 .map(|value| {
                     (
-                        value.i.replace("_", ""),
+                        value.symbol,
                         PricePoint {
-                            price: value.a,
-                            volume: value.v,
+                            price: value.close,
+                            volume: value.base_vol,
                         },
                     )
                 })
