+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/bybit.rs
@@ -4,10 +4,9 @@ use futures::{future::LocalBoxFuture, FutureExt};
 use serde::Deserialize;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
+
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
 
 //TODO(adikov): Include all the needed fields form the response like volume.
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
