+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/gemini.rs
@@ -10,10 +10,9 @@ use serde::Deserialize;
 use serde_json::Value;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
+
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
 
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
 pub struct GeminiPriceResponse {
