+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/kraken.rs
@@ -7,10 +7,9 @@ use futures::{future::LocalBoxFuture, FutureExt};
 use serde::{Deserialize, Deserializer};
 use serde_json::Value;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
+
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
 
 fn as_f64_vec<'de, D>(deserializer: D) -> Result<Vec<f64>, D::Error>
 where
