+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/mexc.rs
@@ -1,48 +1,39 @@
 use anyhow::Result;
 use futures::{future::LocalBoxFuture, FutureExt};
 
-use itertools::Itertools;
 use serde::Deserialize;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
 
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
+
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BinancePriceData {
+pub struct MEXCPriceData {
     pub symbol: String,
     #[serde(deserialize_with = "as_f64")]
+    #[serde(rename = "lastPrice")]
     pub last_price: f64,
     #[serde(deserialize_with = "as_f64")]
     pub volume: f64,
 }
 
-type BinancePriceResponse = Vec<BinancePriceData>;
+type MEXCPriceResponse = Vec<MEXCPriceData>;
 
-pub struct BinancePriceFetcher<'a> {
-    pub symbols: &'a [String],
-}
+pub struct MEXCPriceFetcher;
 
-impl<'a> PricesFetcher<'a> for BinancePriceFetcher<'a> {
-    const NAME: &'static str = "Binance";
+impl PricesFetcher<'_> for MEXCPriceFetcher {
+    const NAME: &'static str = "MEXC";
 
-    fn new(symbols: &'a [String], _api_key: Option<&'a str>) -> Self {
-        Self { symbols }
+    fn new(_symbols: &[String], _api_key: Option<&str>) -> Self {
+        Self
     }
 
     fn fetch(&self) -> LocalBoxFuture<Result<PairPriceData>> {
         async {
-            let req_symbols = format!(
-                "[{}]",
-                self.symbols.iter().map(|s| format!("\"{s}\"")).join(",")
-            );
-
-            let response = http_get_json::<BinancePriceResponse>(
-                "https://api1.binance.com/api/v3/ticker/24hr",
-                Some(&[("symbols", req_symbols.as_str())]),
+            let response = http_get_json::<MEXCPriceResponse>(
+                "https://api.mexc.com/api/v3/ticker/24hr",
+                None,
                 None,
             )
             .await?;
