+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/okx.rs
@@ -4,40 +4,40 @@ use futures::{future::LocalBoxFuture, FutureExt};
 use serde::Deserialize;
 use serde_this_or_that::as_f64;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
 
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
+
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BitgetPriceData {
-    pub symbol: String,
+pub struct OKXTickerData {
+    #[serde(rename = "instId")]
+    pub inst_id: String,
     #[serde(deserialize_with = "as_f64")]
-    pub close: f64,
+    pub last: f64,
     #[serde(deserialize_with = "as_f64")]
-    pub base_vol: f64,
+    pub vol24h: f64,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BitgetPriceResponse {
+pub struct OKXTickerResponse {
     pub code: String,
-    pub data: Vec<BitgetPriceData>,
+    pub data: Vec<OKXTickerData>,
 }
-pub struct BitgetPriceFetcher;
 
-impl PricesFetcher<'_> for BitgetPriceFetcher {
-    const NAME: &'static str = "Bitget";
+pub struct OKXPriceFetcher;
+
+impl PricesFetcher<'_> for OKXPriceFetcher {
+    const NAME: &'static str = "OKX";
 
     fn new(_symbols: &[String], _api_key: Option<&str>) -> Self {
         Self
     }
+
     fn fetch(&self) -> LocalBoxFuture<Result<PairPriceData>> {
         async {
-            let response = http_get_json::<BitgetPriceResponse>(
-                "https://api.bitget.com/api/spot/v1/market/tickers",
-                None,
+            let response = http_get_json::<OKXTickerResponse>(
+                "https://www.okx.com/api/v5/market/tickers",
+                Some(&[("instType", "SPOT")]),
                 None,
             )
             .await?;
@@ -47,10 +47,10 @@ impl PricesFetcher<'_> for BitgetPriceFetcher {
                 .into_iter()
                 .map(|value| {
                     (
-                        value.symbol,
+                        value.inst_id.replace("-", ""),
                         PricePoint {
-                            price: value.close,
-                            volume: value.base_vol,
+                            price: value.last,
+                            volume: value.vol24h,
                         },
                     )
                 })
