+++ libs/data_providers_sdk/src/price_data/fetchers/exchanges/upbit.rs
@@ -3,10 +3,9 @@ use anyhow::Result;
 use futures::{future::LocalBoxFuture, FutureExt};
 use serde::Deserialize;
 
-use blocksense_data_providers_sdk::price_data::traits::prices_fetcher::{
-    PairPriceData, PricePoint, PricesFetcher,
-};
 use blocksense_sdk::http::http_get_json;
+
+use crate::price_data::traits::prices_fetcher::{PairPriceData, PricePoint, PricesFetcher};
 
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
 pub struct UpBitTickerResponseData {
