+++ libs/data_providers_sdk/src/price_data/fetchers/mod.rs
@@ -1,0 +1,1 @@
+pub mod exchanges;
