+++ libs/data_providers_sdk/src/price_data/mod.rs
@@ -1,3 +1,4 @@
+pub mod fetchers;
 pub mod traits;
 pub mod types;
 pub mod wap;
